+++ Src/Private/Get-AbrADTrust.ps1
@@ -111,7 +111,7 @@ function Get-AbrADTrust {
                             if ($Options.EnableDiagrams) {
                                 try {
                                     try {
-                                        $Graph = Get-AbrDiagrammer -DiagramType "Trusts" -DiagramOutput base64 -PSSessionObject $TempPssSession
+                                        $Graph = Get-AbrDiagrammer -DiagramType "Trusts" -DiagramOutput base64 -DomainController $DC
                                     } catch {
                                         Write-PScriboMessage -IsWarning "Domain and Trusts Diagram Graph: $($_.Exception.Message)"
                                     }