+++ Src/Private/Get-AbrDomainSection.ps1
@@ -25,9 +25,8 @@ function Get-AbrDomainSection {
 
     process {
         if ($InfoLevel.Domain -ge 1) {
-            $DomainObj = foreach ($Domain in [string[]]$OrderedDomains) {
-                if ($DomainStatus.Value | Where-Object { $_.Name -eq $Domain -and $_.Status -ne "Offline" }) {
-                    Write-PScriboMessage -IsWarning "Domain $($Domain.ToString().ToUpper()) is offline. Skipping it."
+            $DomainObj = foreach ($Domain in [string[]]($OrderedDomains | Where-Object { $_ -notin $Options.Exclude.Domains })) {
+                if ($Domain -notin $DomainStatus.Value.Name) {
                     if (Get-ValidDCfromDomain -Domain $Domain -DCStatus ([ref]$DCStatus)) {
                         # Define Filter option for Domain variable
                         try {
@@ -88,26 +87,6 @@ function Get-AbrDomainSection {
                                                     }
                                                 }
                                             }
-                                            # Disable function for now, as it is not working properly.
-                                            # if ($HealthCheck.DomainController.Diagnostic) {
-                                            #     try {
-                                            #         $DCDiagObj = foreach ($DC in $DCs) {
-                                            #             if (Get-DCWinRMState -ComputerName $DC) {
-                                            #                 Get-AbrADDCDiag -Domain $Domain -DC $DC
-                                            #             }
-                                            #         }
-                                            #         if ($DCDiagObj) {
-                                            #             Section -Style Heading4 'DC Diagnostic' {
-                                            #                 Paragraph "The following section provides a summary of the Active Directory DC Diagnostic."
-                                            #                 BlankLine
-                                            #                 $DCDiagObj
-                                            #             }
-                                            #         }
-                                            #     } catch {
-                                            #         Write-PScriboMessage -IsWarning "Error: Connecting to remote server $DC failed: WinRM cannot complete the operation. ('DCDiag Information)"
-                                            #         Write-PScriboMessage -IsWarning $_.Exception.Message
-                                            #     }
-                                            # }
                                             try {
                                                 $ADInfraServices = foreach ($DC in $DCs) {
                                                     if (Get-DCWinRMState -ComputerName $DC -DCStatus ([ref]$DCStatus)) {